fix(cloud): restore sidebar desktop and port surfaces
